Move version func to utils
