+++ {{cookiecutter.project_name}}/{{cookiecutter.package_name}}/__init__.py
@@ -1,20 +1,8 @@
 """{{cookiecutter.package_name}}."""
-import importlib.metadata
-from typing import Union
+__all__ = (
+    "__version__",
+    # Add functions and variables you want exposed in `{{cookiecutter.package_name}}.` namespace here
+)
 
-
-def get_version_tuple() -> tuple:
-    def as_integer(x: str) -> Union[int, str]:
-        try:
-            return int(x)
-        except ValueError:
-            return x
-
-    return tuple(
-        as_integer(v)
-        for v in importlib.metadata.version("{{ cookiecutter.package_name }}")
-        .strip()
-        .split(".")
-    )
 
 __version__ = get_version_tuple()
