+++ {{cookiecutter.project_name}}/{{cookiecutter.package_name}}/utils.py
@@ -1,0 +1,17 @@
+import importlib.metadata
+from typing import Union
+
+
+def get_version_tuple() -> tuple:
+    def as_integer(x: str) -> Union[int, str]:
+        try:
+            return int(x)
+        except ValueError:
+            return x
+
+    return tuple(
+        as_integer(v)
+        for v in importlib.metadata.version("{{ cookiecutter.package_name }}")
+        .strip()
+        .split(".")
+    )
